Further optimization of the code

+++ DuEngine/TextureFrameBuffer.h
@@ -16,7 +16,7 @@ public:
 	void reshape(int _width, int _height);
 	vec3 getResolution();
 private:
-	float m_width = 0;
-	float m_height = 0; 
+	int m_width = 0;
+	int m_height = 0;
 };
